requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
django-taggit==0.18.0
django==1.9.4
pytz==2016.3
requests==2.9.1
unidecode==0.4.19
